Several changes to build the addin for Visual Studio for Mac

+++ configure-buildbot.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-#   Copyright 2014 - 2016 Alexander Jochum
+#   Copyright 2014 - 2017 Alexander Jochum
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -37,7 +37,7 @@ fi
 echo "Project is build for MonoDevelop version $MDAPPVERSION"
 echo "MonoDevelop.StyleCop version is $VERSION"
 
-if [ "$MDAPPVERSION" != "6.0" ]; then
+if [ "$MDAPPVERSION" != "6.0" && "$MDAPPVERSION" != "7.0" ]; then
   echo "Removing XS 6.0 ProjectOperationsExtensions.cs file and renaming ProjectOperationsExtensionsPreXS6.cs file appropriately."
   rm ./MonoDevelop.StyleCop/ClassExtensions/ProjectOperationsExtensions.cs
   mv ./MonoDevelop.StyleCop/ClassExtensions/ProjectOperationsExtensionsPreXS6.cs ./MonoDevelop.StyleCop/ClassExtensions/ProjectOperationsExtensions.cs
@@ -79,6 +79,15 @@ echo "Creating files necessary to build the project."
 
 sed "s/INSERT_CSPROJ_VERSION/$VERSION/g" ./MonoDevelop.StyleCop/MonoDevelop.StyleCop.addin.xml.orig > ./MonoDevelop.StyleCop/MonoDevelop.StyleCop.addin.xml
 sed -i.bak "s/INSERT_MAJORAPP_VERSION/$MDAPPVERSION/g" ./MonoDevelop.StyleCop/MonoDevelop.StyleCop.addin.xml
+
+if [ "$MDAPPVERSION" == "7.0" ]; then
+  echo "Patching files for MonoDevelop $MDAPPVERSION"
+  sed -i.bak "s/INSERT_IMMUTABLE_REFERENCE_HERE/<Reference Include=\"System.Collections.Immutable\"><HintPath>INSERT_CSPROJ_MDROOT\\bin\\System.Collections.Immutable.dll</HintPath></Reference>/g" ./MonoDevelop.StyleCop/MonoDevelop.StyleCop-BuildBot.csproj
+  sed -i.bak "s/INSERT_TEXTEDITOR_REFERENCE_HERE/<Reference Include=\"Microsoft.VisualStudio.Text.UI\"><HintPath>INSERT_CSPROJ_MDROOT\\bin\\Microsoft.VisualStudio.Text.UI.dll</HintPath></Reference>/g" ./MonoDevelop.StyleCop/MonoDevelop.StyleCop-BuildBot.csproj
+else
+  sed -i.bak "s/INSERT_IMMUTABLE_REFERENCE_HERE//g" ./MonoDevelop.StyleCop/MonoDevelop.StyleCop-BuildBot.csproj
+  sed -i.bak "s/INSERT_TEXTEDITOR_REFERENCE_HERE/<Reference Include=\"Mono.TextEditor\"><HintPath>INSERT_CSPROJ_MDROOT\\bin\\Mono.TextEditor.dll</HintPath></Reference>/g" ./MonoDevelop.StyleCop/MonoDevelop.StyleCop-BuildBot.csproj
+fi
 
 sed -i.bak "s/INSERT_CSPROJ_VERSION/$VERSION/g" ./MonoDevelop.StyleCop/MonoDevelop.StyleCop-BuildBot.csproj
 sed -i.bak "s/INSERT_TARGET_FRAMEWORKVERSION/$TARGETFRAMEWORKVERSION/g" ./MonoDevelop.StyleCop/MonoDevelop.StyleCop-BuildBot.csproj
